requirements.txt
tensorflow-gpu==1.13.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.1.1
scikit-learn==0.21.3
numpy==1.17.4